added --empty-value option to histogram-median ...
... the median of an empty set is not defined of course, and
the current behaviour of the program is to error, but sometimes
want to treat this case differently to other errors; if the
--empty-value option is passed to the program then its argument
will be printed in place of the median in this case (and the
program exits with success).  Then the caller can check for this
string and handle it appropriately.